requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.2
starlette==0.27.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
unidecode==1.3.7 